agenticai/langchain_agentic_examples/requirements.txt
langchain
langchain-openai
langchain-classic
python-dotenv
pandas
